+++ src/main/java/by/itacademy/jd2/Mk_92_22_Product/service/api/IEssenceService.java
@@ -1,7 +1,6 @@
 package by.itacademy.jd2.Mk_92_22_Product.service.api;
 
 import by.itacademy.jd2.Mk_92_22_Product.core.ProductDTO;
-import by.itacademy.jd2.Mk_92_22_Product.entity.ProductBuilder;
 
 import java.util.List;
 
@@ -9,7 +8,7 @@ public interface IEssenceService<TYPE> {
 
     List<TYPE> get();
 
-    TYPE getId(int id);
+    TYPE getId(Long id);
 
     void validate(ProductDTO item);
