+++ src/main/java/by/itacademy/jd2/Mk_92_22_Product/storage/api/IEssenceStorage.java
@@ -1,6 +1,5 @@
 package by.itacademy.jd2.Mk_92_22_Product.storage.api;
 
-import by.itacademy.jd2.Mk_92_22_Product.entity.Product;
 
 import java.util.List;
 
@@ -8,7 +7,7 @@ public interface IEssenceStorage<TYPE> {
 
     List<TYPE> get();
 
-    TYPE getId (int id);
+    TYPE getId (Long id);
 
     void add (TYPE item);
